Compile most tests on windows

+++ Tests/ForeignTableDmlTest.cpp
@@ -3650,8 +3650,8 @@ TEST_F(SelectQueryTest, CsvArrayEmptyText) {
   sql(result, "SELECT * FROM " + default_table_name + " ORDER BY index;");
   // clang-format off
   assertResultSetEqual({
-                          {1L, array({}), array({"string 1", "string 2"})},
-                          {2L, array({"string 1", "string 2"}), array({})},
+                          {(int64_t)1, array({}), array({"string 1", "string 2"})},
+                          {(int64_t)2, array({"string 1", "string 2"}), array({})},
                       },
     result);
   // clang-format on
@@ -3668,10 +3668,10 @@ TEST_F(SelectQueryTest, CsvArrayNullText) {
   sql(result, "SELECT * FROM " + default_table_name + " ORDER BY index;");
   // clang-format off
   assertResultSetEqual({
-                          {1L, array({Null, Null}), array({Null})},
-                          {2L, array({Null}), array({Null, Null})},
-                          {3L, Null, array({Null, Null})},
-                          {4L, array({Null, Null}), Null},
+                          {(int64_t)1, array({Null, Null}), array({Null})},
+                          {(int64_t)2, array({Null}), array({Null, Null})},
+                          {(int64_t)3, Null, array({Null, Null})},
+                          {(int64_t)4, array({Null, Null}), Null},
                       },
     result);
   // clang-format on
